openAPI/requirements.txt
PyQt5
ccxt
pyjwt
requests
python-telegram-bot
telethon